refactor agreement.js into nevermined srvice
